Modificaciones: (version antes de la conexión con la base de datos)
Form-nuevo-usuario:
Se le agregaron las restricciones a todos los textbox.
Se agregaron todas las verificaciones, controles y funcionalidades.
Se hace uso de la libreria BCrypt para el hasheo de las contraseñas.

+++ proyecto_Villarreal-SanLorenzo/Form-login.cs
@@ -69,7 +69,7 @@ namespace proyecto_Villarreal_SanLorenzo
                                     reader["nombre_usuario"].ToString(),
                                     reader["apellido_usuario"].ToString(),
                                     reader["nombre_rol"].ToString(),
-                                    Convert.ToInt32(reader["telefono"]),
+                                    Convert.ToInt64(reader["telefono"]),
                                     reader["email"].ToString()
                                 );
